+++ src/DataFilters/Grammar/Syntax/OneOfExpression.cs
@@ -7,7 +7,6 @@ namespace DataFilters.Grammar.Syntax
 {
     public class OneOfExpression : FilterExpression, IEquatable<OneOfExpression>
     {
-
         private static readonly ArrayEqualityComparer<FilterExpression> equalityComparer = new ArrayEqualityComparer<FilterExpression>();
 
         public IEnumerable<FilterExpression> Values => _values;