+++ test/DataFilters.TestObjects/Appointment.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace DataFilters.TestObjects
+{
+    public class Appointment
+    {
+        public string Name { get; set; }
+        public DateTimeOffset? Date { get; set; }
+    }
+
+}