+++ test/DataFilters.TestObjects/Weapon.cs
@@ -1,0 +1,10 @@
+namespace DataFilters.TestObjects
+{
+    public class Weapon
+    {
+        public string Name { get; set; }
+
+        public int Level { get; set; }
+    }
+
+}