+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Changed `ConstantExpression` to `ConstantValueExpression` [BREAKING]
+- Moved `FilterExtensions.ToFilter<T>(this StringSegment)` method to `StringExtensions.ToFilter<T>(this StringSegment)` [BREAKING]
+- Moved `FilterExtensions.ToFilter<T>(this string)` method to `StringExtensions.ToFilter<T>(this string)` [BREAKING]
+- Changed subproperty syntax from `property.subproperty` method to `property["subproperty"]` [BREAKING]
+- Added support for filtering over complex type collections.
 
 ## [0.2.2] / 2020-12-05
 - Fixed "contains" parser not working with Guid.